app/api/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-blob
azure-ai-formrecognizer
pypdf
azure-identity
azure-search-documents
azure-ai-formrecognizer
azure-storage-blob
openai
tenacity
